+++ wisdom/include/wisdom/dx12/dx12_views.h
@@ -13,7 +13,6 @@ class Allocation;
 
 namespace wis {
 using DX12FactoryHandle = std::tuple<IDXGIFactory6*>;
-using DX12PipelineHandle = std::tuple<ID3D12PipelineState*>;
 
 using DX12FenceView = std::tuple<ID3D12Fence*>;
 using DX12ShaderView = std::tuple<void*, uint32_t>;
@@ -29,6 +28,7 @@ using DX12DescriptorBufferGPUView = std::tuple<D3D12_GPU_DESCRIPTOR_HANDLE, uint
 using DX12SamplerView = std::tuple<D3D12_CPU_DESCRIPTOR_HANDLE>;
 using DX12ShaderResourceView = std::tuple<D3D12_CPU_DESCRIPTOR_HANDLE>;
 using DX12MemoryView = std::tuple<D3D12MA::Allocator*, D3D12MA::Allocation*>;
+using DX12PipelineView = std::tuple<ID3D12PipelineState*>;
 } // namespace wis
 
 #endif // !WIS_DX12_VIEWS_H